refactor: modularize clustering
